Many updates GH-271

+++ src/map/map/events/timed_drain.rs
@@ -47,7 +47,7 @@ impl Map {
 
             let damage = (character.max_hp as f32 * SETTINGS.world.drain_hp_damage).floor() as i32;
             let damage = cmp::min(damage, character.hp - 1);
-            let damage = cmp::max(damage, 0) as i32;
+            let damage = cmp::max(damage, 0);
 
             character.hp -= damage;
             damage_list.push(damage);
@@ -127,7 +127,7 @@ impl Map {
 
             let damage = (character.max_tp as f32 * SETTINGS.world.drain_tp_damage).floor() as i32;
             let damage = cmp::min(damage, character.tp - 1);
-            let damage = cmp::max(damage, 0) as i32;
+            let damage = cmp::max(damage, 0);
 
             character.tp -= damage;
